src/libical/icalrecur.c - fix a sign-conversion compile warning
Fixes:
```
recur.c:2758:31: error: conversion to ‘int’ from ‘long unsigned int’
may change the sign of the result [-Werror=sign-conversion]
 2758 |                 int new_val = is_limiting
```

+++ src/libical/icalrecur.c
@@ -2757,7 +2757,7 @@ static void expand_by_day(icalrecur_iterator *impl, int year,
             if (valid) {
                 int new_val = is_limiting
                                   /* "Filter" the year days bitmask with the bydays bitmask */
-                                  ? daysmask_getbit(bydays, day + doy_offset)
+                                  ? (int)daysmask_getbit(bydays, day + doy_offset)
                                   /* Add each BYDAY to the year days bitmask */
                                   : 1;
